# Log properties and items on ProjectEvaluationFinished