… src/Build/BackEnd/BuildManager/BuildManager.cs
… src/Build/BackEnd/Components/Logging/EvaluationLoggingContext.cs
… src/Build/BackEnd/Components/Logging/EventSourceSink.cs
… src/Build/BackEnd/Components/Logging/ILoggingService.cs
… src/Build/BackEnd/Components/Logging/LoggingService.cs
… src/Build/BackEnd/Components/Logging/LoggingServiceLogMethods.cs
… src/Build/BackEnd/Components/Logging/ProjectLoggingContext.cs
… src/Build/BackEnd/Node/LoggingNodeConfiguration.cs
… src/Build/BackEnd/Node/OutOfProcNode.cs
… src/Build/Collections/ArrayDictionary.cs
… src/Build/Collections/ItemDictionary.cs
… src/Build/Collections/PropertyDictionary.cs
… src/Build/Definition/ProjectCollection.cs
… src/Build/Evaluation/Evaluator.cs
… src/Build/Instance/ProjectInstance.cs
… src/Build/Logging/BaseConsoleLogger.cs
… src/Build/Logging/BinaryLogger/BinaryLogger.cs
… src/Build/Logging/BinaryLogger/BuildEventArgsReader.cs
… src/Build/Logging/BinaryLogger/BuildEventArgsWriter.cs
… src/Build/Logging/ParallelLogger/ParallelConsoleLogger.cs
… src/Build/Logging/SerialConsoleLogger.cs
… src/Build/Utilities/Utilities.cs
… src/Framework/BuildEventArgs.cs
… src/Framework/IEventSource4.cs
… src/Framework/ProjectEvaluationFinishedEventArgs.cs
… src/Framework/ProjectEvaluationStartedEventArgs.cs
… src/MSBuild/MSBuild.csproj
… src/Shared/LogMessagePacketBase.cs
… src/Shared/Traits.cs
… src/Utilities/MuxLogger.cs