@KirillOsenkov 2021-03-20
@mmitche it would be really good if you could take this one for a spin and get some stats on build perf and binlog size for a really large build.

You can use `msbuild -bl -c Release /p:CreateBootstrap=true` to produce an MSBuild layout and then use `C:\MSBuild\artifacts\bin\bootstrap\net472\MSBuild\Current\Bin\MSBuild.exe` to build.