@KirillOsenkov 2021-03-22
I have updated the binlog viewer for the latest format and added some affordances to make it easier to look up properties and items now that they're logged at evaluation and not under the project:

![image](https://user-images.githubusercontent.com/679326/112044293-e2703080-8b06-11eb-9d57-675be814ce1e.png)

![image](https://user-images.githubusercontent.com/679326/112044008-886f6b00-8b06-11eb-89e5-84ca72305853.png)

![image](https://user-images.githubusercontent.com/679326/112044043-91f8d300-8b06-11eb-9a4c-356a0acf7497.png)

![image](https://user-images.githubusercontent.com/679326/112044124-aa68ed80-8b06-11eb-8634-ffcc73c74bd2.png)

![image](https://user-images.githubusercontent.com/679326/112044262-d5534180-8b06-11eb-81a4-05b65330122f.png)