@KirillOsenkov 2021-03-24
I got some stats for building Roslyn.

Binlog size:
16.9: 122MB
main: 14.7MB
PR 6287: 15.6MB

Against expectations, the binlog size is increased slightly with this PR, but at least we log all properties and items for each project, not just a small part. You can tell whether it has properties and items by the presence of the target framework label:
![image](https://user-images.githubusercontent.com/679326/112381723-25b5d500-8ca8-11eb-8f31-a3d94dab360f.png)

The build time seemed very close for this PR vs. main.