@KirillOsenkov 2021-03-30
And yes, `protected internal` means protected OR internal (union). So everyone in the same assembly + all assemblies that this assembly has InternalsVisibleTo to + all derived types.

C# recently introduced `private protected` to mean protected AND internal (intersection). Only types in the same assembly + IVT which derive from this one.